% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_DEP.R
\docType{data}
\name{boundaries_DEP}
\alias{boundaries_DEP}
\title{Departmental boundaries of Peru}
\format{
dataframe
\describe{
  \item{long}{dbl Longitude of the departmental boundary}
  \item{lat}{dbl Latitude of the departmental boundary}
  \item{COD_DEPARTAMENTO}{chr Department Code}
  \item{DEPARTAMENTO}{chr Department name}
  \item{group}{dbl Group identifier on map}
}
}
\usage{
boundaries_DEP
}
\description{
Departmental boundaries of Peru
}
\keyword{datasets}
