% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_DIST.R
\docType{data}
\name{boundaries_DIST}
\alias{boundaries_DIST}
\title{District boundaries of Peru}
\format{
dataframe
\describe{
  \item{long}{dbl Longitude of the district boundary}
  \item{lat}{dbl Latitude of the district boundary}
  \item{COD_DISTRITO}{chr District Code}
  \item{DEPARTAMENTO}{chr Department name}
  \item{PROVINCIA}{chr Province name}
  \item{DISTRITO}{chr District name}
  \item{group}{dbl Group identifier on map}
}
}
\usage{
boundaries_DIST
}
\description{
District boundaries of Peru
}
\keyword{datasets}
