% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_REG.R
\docType{data}
\name{boundaries_REG}
\alias{boundaries_REG}
\title{Regional boundaries of Peru}
\format{
dataframe
\describe{
  \item{long}{dbl longitud}
  \item{lat}{dbl latitud}
  \item{COD_REGION}{chr Código de la región}
  \item{REGION}{chr Nombre de la región}
  \item{group}{fct identificador de grupo en mapa}
}
}
\usage{
boundaries_REG
}
\description{
Regional boundaries of Peru
}
\keyword{datasets}
