% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids_DEP.R
\docType{data}
\name{centroids_DEP}
\alias{centroids_DEP}
\title{Departmental centroids of Peru}
\format{
dataframe
\describe{
  \item{long}{dbl Longitude of the departmental centroid}
  \item{lat}{dbl Latitude of the departmental centroid}
  \item{COD_DEPARTAMENTO}{chr Department Code}
  \item{DEPARTAMENTO}{chr Department name}
  \item{geometry}{GEOMETRY Geometric object}
}
}
\usage{
centroids_DEP
}
\description{
Departmental centroids of Peru
}
\keyword{datasets}
