% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids_DIST.R
\docType{data}
\name{centroids_DIST}
\alias{centroids_DIST}
\title{District centroids of Peru}
\format{
dataframe
\describe{
  \item{long}{dbl Longitude of the regional centroid}
  \item{lat}{dbl Latitude of the regional centroid}
  \item{COD_DISTRITO}{chr District Code}
  \item{DEPARTAMENTO}{chr Department name}
  \item{PROVINCIA}{chr Province name}
  \item{DISTRITO}{chr District name}
  \item{geometry}{GEOMETRY Geometric object}
}
}
\usage{
centroids_DIST
}
\description{
District centroids of Peru
}
\keyword{datasets}
