% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids_REG.R
\docType{data}
\name{centroids_REG}
\alias{centroids_REG}
\title{Regional centroids of Peru}
\format{
dataframe
\describe{
  \item{long}{dbl Longitude of the regional centroid}
  \item{lat}{dbl Latitude of the regional centroid}
  \item{COD_REGION}{chr Region Code}
  \item{REGION}{chr Region name}
  \item{geometry}{GEOMETRY Geometric object}
}
}
\usage{
centroids_REG
}
\description{
Regional centroids of Peru
}
\keyword{datasets}
