% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_REG.R
\name{get_REG}
\alias{get_REG}
\title{Regional boundaries of Peru}
\usage{
get_REG(x = 1)
}
\arguments{
\item{x}{indicates whether you want to obtain the regional boundaries or the centroids. Choose 1 or leave this parameter empty if you want to obtain the regional boundaries. Choose 2 if you want to obtain the centroids of each region.}
}
\description{
Load the dataset of the geographic limits of the regions of Peru.
}
\details{
Officially there is no codification for regions, only for departments. Therefore, the codes 150100 for Metropolitan Lima and 159900 for Lima Provinces should be taken as a reference.
}
\examples{
# regional boundaries of Peru
get_REG()
get_REG(1)

# centroids of each region of Peru
get_REG(2)

# use regional boundaries of Peru in a map with ggplot2
boundaries <- get_REG(1)
library(ggplot2)
ggplot(data=boundaries)+
aes(x=long, y=lat, group=group)+
geom_polygon(aes(fill=REGION))+
coord_equal()+
geom_path(color="black")

# use regional boundaries of Peru and centroids in a map with ggplot2
boundaries <- get_REG(1)
centroids <- get_REG(2)
library(ggplot2)
ggplot(data=boundaries)+
aes(x=long, y=lat, group=group)+
geom_polygon(aes(fill=REGION))+
coord_equal()+
geom_path(color="black")+
geom_text(data=centroids, aes(long, lat, group=NULL, label=REGION), size=2)

}
