% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_DEP.R
\docType{data}
\name{map_DEP}
\alias{map_DEP}
\title{Geographic information of the departments of Peru}
\format{
dataframe
\describe{
  \item{COD_DEPARTAMENTO}{chr Department Code}
  \item{DEPARTAMENTO}{chr Department name}
  \item{coords_x}{dbl Longitude of the centroid of the department}
  \item{coords_y}{dbl Latitude of the centroid of the department}
  \item{geometry}{MULTIPOLYGON Geometric object}
}
}
\usage{
map_DEP
}
\description{
Geographic information of the departments of Peru
}
\keyword{datasets}
