% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_DIST.R
\docType{data}
\name{map_DIST}
\alias{map_DIST}
\title{Geographic information of the districts of Peru}
\format{
dataframe
\describe{
  \item{COD_REGION}{chr Region Code}
  \item{COD_DEPARTAMENTO}{chr Department Code}
  \item{COD_PROVINCIA}{chr Province Code}
  \item{COD_DISTRITO}{chr District Code}
  \item{REGION}{chr Region name}
  \item{DEPARTAMENTO}{chr Department name}
  \item{PROVINCIA}{chr Province name}
  \item{DISTRITO}{chr District name}
  \item{coords_x}{dbl Longitude of the centroid of the district}
  \item{coords_y}{dbl Latitude of the centroid of the district}
  \item{geometry}{MULTIPOLYGON Geometric object}
}
}
\usage{
map_DIST
}
\description{
Geographic information of the districts of Peru
}
\keyword{datasets}
