% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_PROV.R
\docType{data}
\name{map_PROV}
\alias{map_PROV}
\title{Geographic information of the provinces of Peru}
\format{
dataframe
\describe{
  \item{COD_PROVINCIA}{chr Province Code}
  \item{DEPARTAMENTO}{chr Department name}
  \item{PROVINCIA}{chr Province name}
  \item{coords_x}{dbl Longitude of the centroid of the province}
  \item{coords_y}{dbl Latitude of the centroid of the province}
  \item{geometry}{MULTIPOLYGON Geometric object}
}
}
\usage{
map_PROV
}
\description{
Geographic information of the provinces of Peru
}
\keyword{datasets}
