% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_REG.R
\docType{data}
\name{map_REG}
\alias{map_REG}
\title{Geographic information of the regions of Peru}
\format{
dataframe
\describe{
  \item{COD_REGION}{chr Region Code}
  \item{REGION}{chr Region name}
  \item{coords_x}{dbl Latitude of the centroid of the region}
  \item{coords_y}{dbl Longitude of the centroid of the region}
  \item{geometry}{MULTIPOLYGON Geometric object}
}
}
\usage{
map_REG
}
\description{
Geographic information of the regions of Peru
}
\keyword{datasets}
