% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_DEP.R
\docType{data}
\name{boundaries_DEP}
\alias{boundaries_DEP}
\title{Departmental boundaries of Peru}
\format{
dataframe
\describe{
  \item{long}{dbl longitud del límite departamental}
  \item{lat}{dbl latitud del límite departamental}
  \item{COD_DEPARTAMENTO}{chr Código del departamento}
  \item{DEPARTAMENTO}{chr Nombre del departamento}
  \item{group}{fct identificador de grupo en mapa}
}
}
\usage{
boundaries_DEP
}
\description{
Departmental boundaries of Peru
}
\keyword{datasets}
