% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_REG.R
\docType{data}
\name{boundaries_REG}
\alias{boundaries_REG}
\title{Regional boundaries of Peru}
\format{
dataframe
\describe{
  \item{long}{dbl Longitude of the regional boundary}
  \item{lat}{dbl Latitude of the regional boundary}
  \item{COD_REGION}{chr Region Code}
  \item{REGION}{chr Region name}
  \item{group}{dbl Group identifier on map}
}
}
\usage{
boundaries_REG
}
\description{
Regional boundaries of Peru
}
\keyword{datasets}
