% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids_DEP.R
\docType{data}
\name{centroids_DEP}
\alias{centroids_DEP}
\title{Departmental centroids of Peru}
\format{
dataframe
\describe{
  \item{long}{dbl longitud del centroide}
  \item{lat}{dbl latitud del centroide}
  \item{COD_DEPARTAMENTO}{chr Código del departamento}
  \item{DEPARTAMENTO}{chr Nombre del departamento}
}
}
\usage{
centroids_DEP
}
\description{
Departmental centroids of Peru
}
\keyword{datasets}
