% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids_REG.R
\docType{data}
\name{centroids_REG}
\alias{centroids_REG}
\title{Regional centroids of Peru}
\format{
dataframe
\describe{
  \item{long}{dbl longitud del centroide}
  \item{lat}{dbl latitud del centroide}
  \item{COD_REGION}{chr Código de la regíón}
  \item{REGION}{chr Nombre de la regíón}
}
}
\usage{
centroids_REG
}
\description{
Regional centroids of Peru
}
\keyword{datasets}
