% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsPERU-package.R
\docType{package}
\name{mapsPERU-package}
\alias{mapsPERU}
\alias{mapsPERU-package}
\title{mapsPERU: Maps of Peru}
\description{
Information of the centroids and geographical limits of the regions, departments and districts of Peru.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/musajajorge/mapsPERU}
}

}
\author{
\strong{Maintainer}: Jorge L. C. Musaja \email{musajajorge@gmail.com}

}
\keyword{internal}
